[nodejs-review] review-10 fixed. move config file out of src dir
