+++ projects/koajs-proj-conf/src/routes.js
@@ -1,6 +1,6 @@
 const Router = require('koa-router');
 
-const config = require('./config');
+const config = require('../config');
 const controllers = require('./controllers');
 const {
     bookItemValidators,
